+++ pydantic_resolve/__init__.py
@@ -1,3 +1,6 @@
+from .utils.class_util import ensure_subset
+from .utils.dataloader import build_list, build_object, copy_dataloader_kls
+from .utils.conversion import mapper
 from .exceptions import (
     ResolverTargetAttrNotFound,
     DataloaderDependCantBeResolved,
@@ -6,14 +9,8 @@ from .exceptions import (
     GlobalLoaderFieldOverlappedError)
 from .resolver import Resolver
 from .core import LoaderDepend, Collector, ICollector
-from .util import (
-    build_list,
-    build_object,
-    mapper,
-    ensure_subset,
-    output,
-    model_config,
-    copy_dataloader_kls)
+from .utils.openapi import (
+    model_config)
 
 
 __all__ = [
@@ -31,7 +28,6 @@ __all__ = [
     'build_object',
     'mapper',
     'ensure_subset',
-    'output',
     'model_config',
     'copy_dataloader_kls',
 ]
