+++ pydantic_resolve/utils/params.py
@@ -1,0 +1,12 @@
+from pydantic_resolve.exceptions import GlobalLoaderFieldOverlappedError
+
+
+from typing import Any, Dict
+
+
+def merge_dicts(a: Dict[str, Any], b: Dict[str, Any]):
+    overlap = set(a.keys()) & set(b.keys())
+    if overlap:
+        raise GlobalLoaderFieldOverlappedError(f'loader_params and global_loader_param have duplicated key(s): {",".join(overlap)}')
+    else:
+        return {**a, **b}
