+++ tests/pydantic_v2/utils/test_2_ensure_subset.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass
 from pydantic import BaseModel
-from pydantic_resolve import util
+from pydantic_resolve.utils import class_util
 import pytest
 
 def test_ensure_subset():
@@ -13,11 +13,11 @@ def test_ensure_subset():
         a: str
         b: int
     
-    @util.ensure_subset(Base)
+    @class_util.ensure_subset(Base)
     class ChildA(BaseModel):
         a: str
 
-    @util.ensure_subset(Base)
+    @class_util.ensure_subset(Base)
     class ChildB(BaseModel):
         a: str
         c: int = 0
@@ -25,20 +25,20 @@ def test_ensure_subset():
             return 21
 
     with pytest.raises(AttributeError):
-        @util.ensure_subset(Base)
+        @class_util.ensure_subset(Base)
         class ChildC(BaseModel):
             a: str
             b: str
 
     with pytest.raises(AttributeError):
-        @util.ensure_subset(Base)
+        @class_util.ensure_subset(Base)
         class ChildD(BaseModel):
             a: str
             b: int
             c: int
 
     with pytest.raises(AssertionError):
-        @util.ensure_subset(Base)
+        @class_util.ensure_subset(Base)
         @dataclass
         class ChildE():
             a: str
@@ -46,7 +46,7 @@ def test_ensure_subset():
             c: int
 
     with pytest.raises(AssertionError):
-        @util.ensure_subset(BaseX)
+        @class_util.ensure_subset(BaseX)
         @dataclass
         class ChildF():
             a: str
